+++ requirements.txt
@@ -1,0 +1,12 @@
+ipython >=4
+jupyter_client
+jupyter_core
+notebook
+nbconvert
+nbformat
+python
+traitlets
+psutil
+traitlets
+pyaml
+
